+++ src/types.rs
@@ -104,7 +104,9 @@ pub struct SequencesSketch{
     pub c: usize,
     pub k: usize,
     pub file_name: String,
-    pub paired: bool
+    pub sample_name: Option<String>,
+    pub paired: bool,
+    pub mean_read_length: f64,
 }
 
 //Encoding kmer_counts as vec speeds up serialize/deserialize by
@@ -115,7 +117,9 @@ pub struct SequencesSketchEncode{
     pub c: usize,
     pub k: usize,
     pub file_name: String,
-    pub paired: bool
+    pub sample_name: Option<String>,
+    pub paired: bool,
+    pub mean_read_length: f64,
 }
 
 impl SequencesSketchEncode{
@@ -124,13 +128,13 @@ impl SequencesSketchEncode{
         for (key,val) in sketch.kmer_counts.into_iter(){
             vec_map.push((key,val));
         }
-        return SequencesSketchEncode{kmer_counts: vec_map, file_name: sketch.file_name, c: sketch.c, k: sketch.k, paired: sketch.paired};
+        return SequencesSketchEncode{kmer_counts: vec_map, file_name: sketch.file_name, c: sketch.c, k: sketch.k, paired: sketch.paired, mean_read_length: sketch.mean_read_length, sample_name: sketch.sample_name };
     }
 }
 
 impl SequencesSketch{
-    pub fn new(file_name: String, c: usize, k: usize, paired: bool) -> SequencesSketch{
-        return SequencesSketch{kmer_counts : HashMap::default(), file_name, c, k, paired}
+    pub fn new(file_name: String, c: usize, k: usize, paired: bool, sample_name: Option<String>, mean_read_length: f64) -> SequencesSketch{
+        return SequencesSketch{kmer_counts : HashMap::default(), file_name, c, k, paired, sample_name, mean_read_length}
     }
     pub fn from_enc(sketch: SequencesSketchEncode) -> SequencesSketch{
         let mut new_map = FxHashMap::default();
@@ -138,7 +142,7 @@ impl SequencesSketch{
         for item in sketch.kmer_counts.into_iter(){
             new_map.insert(item.0, item.1);
         }
-        return SequencesSketch{kmer_counts: new_map, file_name: sketch.file_name, c: sketch.c, k: sketch.k, paired: sketch.paired};
+        return SequencesSketch{kmer_counts: new_map, file_name: sketch.file_name, c: sketch.c, k: sketch.k, paired: sketch.paired, mean_read_length: sketch.mean_read_length, sample_name: sketch.sample_name};
     }
 }